added descriptive asserts and added utils file
