+++ pgmpy/utils/sets.py
@@ -1,0 +1,38 @@
+from collections import Iterable
+from itertools import combinations, chain
+
+
+def _variable_or_iterable_to_set(x):
+    """
+    Convert variable, set, or iterable x to a frozenset.
+
+    If x is None, returns the empty set.
+
+    Parameters
+    ---------
+    x : None, str or Iterable[str]
+    """
+    if x is None:
+        return frozenset([])
+
+    if isinstance(x, str):
+        return frozenset([x])
+
+    if isinstance(x, set):
+        return frozenset(x)
+
+    if not isinstance(x, Iterable) or not all(isinstance(xx, str) for xx in x):
+        raise ValueError(
+            "{} is expected to be either a string, set of strings, or an iterable of strings"
+            .format(x))
+
+    return frozenset(x)
+
+
+def _powerset(iterable):
+    """
+    https://docs.python.org/3/library/itertools.html#recipes
+    powerset([1,2,3]) --> () (1,) (2,) (3,) (1,2) (1,3) (2,3) (1,2,3)
+    """
+    s = list(iterable)
+    return chain.from_iterable(combinations(s, r) for r in range(len(s) + 1))
